% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPH-fun.R
\name{adfgvx.cipher}
\alias{adfgvx.cipher}
\title{ADFGVX Cipher.}
\usage{
adfgvx.cipher(ms, key, code)
}
\arguments{
\item{ms}{The message to be encrypted/decrypted in a single character object. The Raw message can be written with upper cases and spaces and numbers can be used. The encrypted message needs a space between each five-letter symbol.}

\item{key}{A single word, necessarily one where no letters are repeated i.e. "Airjet", which determines the unique pattern in which the message is encrypted or decrypted.}

\item{code}{It refers to the transformation that will be made to the message: "encrypt" and "decrypt".}
}
\value{
For the encrypted message, returns a list of two: "Encrypted Message", which is the set of five-letter transformation of the Raw message; and "Null (characters) Values Added", which is the number of characters added into the message - that is because additional letters need to be added to complete the Transposition matrix. For the decrypted message, returns a single character object with the decrypted message.
}
\description{
A cipher with two transformations: Substitution and Transposition. The original message is replaced by the values inside the ADFGVX square, and then, using a given keyword, the transformed message is rearranged, obtaining a unique set of code.
}
\examples{
adfgvx.cipher(ms = "Hello World 1", key = "Post", code = "encrypt")

adfgvx.cipher(ms = "DXXXA XXAGA DGAXD FAAGD AXGAD AXDFD", key = "Air", code = "decrypt")
}
