% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPH-fun.R
\name{beaufort.cipher}
\alias{beaufort.cipher}
\title{Beaufort Cipher.}
\usage{
beaufort.cipher(ms, key, code)
}
\arguments{
\item{ms}{The message to be encrypted/decrypted in a single character object. The Raw and encrypted message can be written with spaces and upper cases, but no numbers.}

\item{key}{A single letter character object (in upper or lower case) which indicates the key in which the message will be encrypted or decrypted.}

\item{code}{It refers to the transformation that will be made to the message: "encrypt" and "decrypt".}
}
\value{
A character object with the transformed message, using the key to locate in the beaufort code the corresponding letter.
}
\description{
Substitution cipher, using the Beaufort square and a given letter as a Key.
}
\examples{
beaufort.cipher(ms = "Hello World", key = "Z", code = "encrypt")

beaufort.cipher(ms = "FIBBYQYVBJ", key = "m", code = "decrypt")
}
