% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPH-fun.R
\name{caesar.cipher}
\alias{caesar.cipher}
\title{Caesar Cipher.}
\usage{
caesar.cipher(ms, pos.shift, code)
}
\arguments{
\item{ms}{The message to be encrypted/decrypted in a single character object. The Raw and encrypted message can be written with upper cases and spaces, but no numbers.}

\item{pos.shift}{Indicates the position the alphabet will be shifted. Only accepts an integer between 1:26.}

\item{code}{It refers to the transformation that will be made to the message: "encrypt" and "decrypt".}
}
\value{
A character object with the transformed message, using a position shift inside the alphabet to encrypt/decrypt the message.
}
\description{
Transforms by substitution a given message using a position shift of the alphabet, substituting each letter in the Raw message for a the letter that is a fixed number of positions ahead.
}
\examples{
caesar.cipher(ms = "Hello World", pos.shift = 1, code = "encrypt")

caesar.cipher(ms = "jgnnqyqtnf", pos.shift = 2, code = "decrypt")
}
