% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPH.R
\docType{package}
\name{CIPH}
\alias{CIPH}
\title{CIPH: A package for encrypting and deciphering hidden messages.}
\description{
The CIPH package contains a series of simple functions that can encrypt or crack with multiple cipher codes, each with its unique qualities. The available ciphers are:
-Numeric Cipher
-Caesar Cipher
-Polybius Cipher
-ADFGVX Cipher
-Bacon Cipher
-Beaufort Cipher
}
