% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPH-fun.R
\name{num.cipher}
\alias{num.cipher}
\title{Numeric Cipher.}
\usage{
num.cipher(ms, code)
}
\arguments{
\item{ms}{The message to be encrypted/decrypted in a single character object. The Raw message can be written with upper cases and spaces, but no numbers. The encrypted message has to be written with numbers, each separated by a single space.}

\item{code}{It refers to the transformation that will be made to the message: "encrypt" and "decrypt".}
}
\value{
A character object with the transformed message, using the replacement of letters to numbers or vice versa.
}
\description{
Transforms by substitution a given message using the number in each letter of the alphabet (a = 1, b = 2, c = 3, ...).
}
\examples{
num.cipher(ms = "Hello World", code = "encrypt")

num.cipher(ms = "8 5 12 12 15 23 15 18 12 4", code = "decrypt")
}
