% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIPH-fun.R
\name{polyb.cipher}
\alias{polyb.cipher}
\title{Polybius Cipher.}
\usage{
polyb.cipher(ms, code)
}
\arguments{
\item{ms}{The message to be encrypted/decrypted in a single character object. The Raw message can be written with upper cases and spaces, but no numbers. The transformed message has to be written with numbers, each separated by a single space.}

\item{code}{It refers to the transformation that will be made to the message: "encrypt" and "decrypt".}
}
\value{
A character object with the transformed message, using the unique matrix of the polybius cipher. The letters "i" and "j" have the same symbol.
}
\description{
Using the Polybius square, substituting each letter for a coordinate it holds inside the square.
}
\examples{
polyb.cipher(ms = "Hello World", code = "encrypt")

polyb.cipher(ms = "23 15 31 31 34 52 34 42 31 14", code = "decrypt")
}
